requirements.txt
opentelemetry-distro[otlp]
opentelemetry-instrumentation